refactor: #89 물품 자료형 int로 수정

+++ src/main/java/com/ajou/hertz/domain/instrument/repository/InstrumentRepository.java
@@ -7,6 +7,6 @@ import com.ajou.hertz.domain.instrument.entity.Instrument;
 
 public interface InstrumentRepository extends JpaRepository<Instrument, Long>, InstrumentRepositoryCustom {
 
-	long countBySellerIdAndProgressStatus(Long sellerid, InstrumentProgressStatus progressStatus);
+	int countBySellerIdAndProgressStatus(Long sellerid, InstrumentProgressStatus progressStatus);
 
 }
